Move StoryComponentCache to its own file
