+++ SimplifiedIgStories/Main/SimplifiedIgStoriesApp.swift
@@ -26,24 +26,6 @@ final class AppComponentsFactory {
     )
 }
 
-final class StoryComponentCache<T> {
-    typealias StoryId = Int
-    
-    private var cache = [StoryId: T]()
-    
-    func save(_ component: T, for storyId: StoryId) {
-        cache[storyId] = component
-    }
-    
-    func getComponent(for storyId: StoryId) -> T? {
-        cache[storyId]
-    }
-    
-    func removeComponent(for storyId: StoryId) {
-        cache[storyId] = nil
-    }
-}
-
 @main
 struct SimplifiedIgStoriesApp: App {
     private let factory = AppComponentsFactory()
